use -nologo option for VC Perl (nmake) builds
Visual C++ builds of Perl normally use nmake (other builds can too though
I think this is fairly uncommon today).  Unfortunately nmake outputs a
copyright message by default before doing anything and this can cause
the tests for [DynamicPrereqs] to fail as they rely on silent make output.
This patch adds the -nologo option in the helper so individual tests do
not have to set it.

This seems to be a good way to address rt#131185.